file: move stat_data definition to file-types.hh
